chore(OBS-2182): rename huawei_vrp driver to vrp
Rename custom_napalm/huawei_vrp.py → vrp.py and the corresponding test
directory tests/custom_drivers/huawei_vrp/ → vrp/ for a shorter, OS-
agnostic driver name. The NAPALM driver short-name changes from
`huawei_vrp` to `vrp` accordingly.
