+++ device-discovery/custom_napalm/__init__.py
@@ -20,7 +20,6 @@ from custom_napalm.ftd_ssh import FTDSSHDriver
 from custom_napalm.ftos import FTOSDriver
 from custom_napalm.fxos import FXOSDriver
 from custom_napalm.gaia import GaiaDriver
-from custom_napalm.huawei_vrp import VRPDriver
 from custom_napalm.ipos import IPOSDriver
 from custom_napalm.netiron import NetIronDriver
 from custom_napalm.panos import PANOSDriver
@@ -37,6 +36,7 @@ from custom_napalm.sros import SROSDriver
 from custom_napalm.sros_ssh import SROSSSHDriver
 from custom_napalm.unifiswitch import UniFiSwitchDriver
 from custom_napalm.viptela_ssh import ViptelaSSHDriver
+from custom_napalm.vrp import VRPDriver
 from custom_napalm.vsp import VSPDriver
 from custom_napalm.wlc import WLCDriver
