+++ device-discovery/custom_napalm/vrp.py
