+++ device-discovery/tests/custom_drivers/vrp/mock_data/test_get_config/normal/display_current-configuration.txt
