+++ device-discovery/tests/custom_drivers/vrp/mock_data/test_get_config_sanitized/normal/display_saved-configuration.txt
