+++ device-discovery/tests/custom_drivers/vrp/mock_data/test_get_facts/normal/display_version.txt
