+++ device-discovery/tests/custom_drivers/vrp/mock_data/test_get_interfaces/normal/display_interface.txt
