+++ device-discovery/tests/custom_drivers/vrp/mock_data/test_get_interfaces_ip/normal/display_ipv6_interface.txt
