+++ device-discovery/tests/custom_drivers/vrp/mock_data/test_get_vlans/normal/display_vlan_brief.txt
